+++ src/plugins/include/helpers.hpp
@@ -9,7 +9,7 @@
 #define BASEAMPLITUDE 32000
 
 //duty between 0 and 1 and represents portion of time up
-float GetSquareWave(float hz, float duty, float amplitude, float phaseDiff, float time);
+float GetSquareWave(float hz, float duty, float amplitude, float phaseDiff, double time);
 
 float GetBeatTime(float bpm, int beat, int numerator, int denom);
 float GetTimeBetweenBeats(float bpm, int numerator, int denom);
