+++ src/plugins/include/sequence.hpp
@@ -20,7 +20,7 @@ typedef struct measure {
 
 typedef struct sequenceSample {
     Sample *sample;
-    float startTime;
+    double startTime;
 } SequenceSample;
 
 typedef struct drawnSample {
@@ -29,7 +29,7 @@ typedef struct drawnSample {
 } DrawnSample;
 
 extern bool isClickUsed; //redefinition of the thing in maingui.hpp for use by other seqs
-extern float currentTime; //redefinition of the thing in manigui.hpp for use by other seqs
+extern double currentTime; //redefinition of the thing in manigui.hpp for use by other seqs
 
 enum Edge {None, Bottom, BottomRight, BottomLeft, Left, TopLeft, Top, TopRight, Right, Header };
 
@@ -38,9 +38,9 @@ class Sequence {
     public:
         Sequence(); //empty
         virtual ~Sequence(); //destructor
-        //void Sequence::AddSamplesOfLength(std::shared_ptr<Sample> sample, float startTime, float freqMult, float length, int repetitions = 1, float timeGap = 0);
-        virtual float GetSampleAtTime(float time); //updates sample logic based off time and gets sample, should be in range 0,1
-        virtual SequenceSample * AddSamples(std::unordered_map<std::string, SampleProperty>, float startTime, int repetitions = 1, float timeGap = 0);
+        //void Sequence::AddSamplesOfLength(std::shared_ptr<Sample> sample, double startTime, float freqMult, float length, int repetitions = 1, double timeGap = 0);
+        virtual float GetSampleAtTime(double time); //updates sample logic based off time and gets sample, should be in range 0,1
+        virtual SequenceSample * AddSamples(std::unordered_map<std::string, SampleProperty>, double startTime, int repetitions = 1, double timeGap = 0);
         void DrawWindow(); //draws window
         //void DrawSequence(); //draws sequence to back via maingui
         virtual void Update();
@@ -55,11 +55,11 @@ class Sequence {
         Rectangle GetCurrentWindowPos(); //include header
         Rectangle GetCurrentWindowPaddedPos(); //include side padding for resize
         std::vector<SequenceSample*> GetAllSamples(); // orders samples 
-        std::vector<std::tuple<Measure, float>> GetMeasures(); //float is measure start time
+        std::vector<std::tuple<Measure, double>> GetMeasures(); //float is measure start time
         std::vector<DrawnSample> lastDrawnSamples;
         SequenceSample *selectedSamp = nullptr;
         SequenceSample *ghostSamp = nullptr;
-        std::tuple<Measure, float> *GetMeasureAtTime(float time);
+        std::tuple<Measure, double> *GetMeasureAtTime(double time);
         float seqHeight;
         float seqYPos;
         bool wasSeqResizeSelected;
@@ -73,21 +73,21 @@ class Sequence {
         std::vector<SequenceSample*> activeSamples;
         std::vector<SequenceSample*> samplesToAdd; // low to high (front to back)
         std::vector<SequenceSample*> samplesAdded; // high to low (front to back)
-        void AddSequenceSample(SequenceSample *, float startTime); //ONLY CALL WITH PREPOPULATED SAMPLE
+        void AddSequenceSample(SequenceSample *, double startTime); //ONLY CALL WITH PREPOPULATED SAMPLE
         void UpdateCurrentPos(Rectangle rect); //updates rendertexture too
         bool isWindowResized;
         Edge selectedEdge = Edge::None;
         float GetBeatTime(int measure, float beat); //measures are 1 indexed
-        float GetBeatAtTime(float time); //returns beat in the measure at time
+        float GetBeatAtTime(double time); //returns beat in the measure at time
         void SortSamplesToAdd();
         void SortSamplesAdded();
-        float prevTime = 0;
+        double prevTime = 0;
     private:
         std::string loadedSongPath;
         void AddMeasureToCount(Measure measure);
         void UpdateMeasureTimes();
         //Measure, and time at the start of that measure
-        std::vector<std::tuple<Measure, float>> measures; //don't directly add to this
+        std::vector<std::tuple<Measure, double>> measures; //don't directly add to this
         Rectangle currentPos;
         RenderTexture2D windowTex; 
         void UpdateCurrentPos();
